kafka_ex/requirements.txt
kafka-python==2.0.2
requests==2.28.1
six==1.16.0
python-telegram-bot==13.7


